# backend/src/app.js
const path = require('path');
require('dotenv').config({ path: path.resolve(__dirname, '../.env') });
const express = require('express');
const cors = require('cors');
const morgan = require('morgan');
const { initializeDatabase } = require('./shared/database');
const authMiddleware = require('./shared/middleware/auth');

const app = express();

const allowedOrigins = new Set([
  'https://att.jamfortetech.com',
  process.env.FRONTEND_URL,
].filter(Boolean).map(origin => origin.replace(/\/$/, '')));

if (process.env.NODE_ENV !== 'production') {
  allowedOrigins.add('http://localhost:3000');
  allowedOrigins.add('http://localhost:5173');
}

app.use(cors({
  origin(origin, callback) {
    // Requests without an Origin header (health checks, server-to-server calls)
    // are not browser CORS requests and should remain available.
    if (!origin || allowedOrigins.has(origin.replace(/\/$/, ''))) {
      return callback(null, true);
    }
    const error = new Error(`Origin ${origin} is not allowed by CORS`);
    error.status = 403;
    return callback(error);
  },
  credentials: true,
}));
app.use(express.json());
app.use(morgan('dev'));

const swaggerUi = require('swagger-ui-express');
const swaggerDocument = require('../swagger.json');

app.get('/api/health', (_req, res) => res.json({ status: 'ok' }));

app.get('/api/test-email', async (_req, res) => {
  const sgMail = require('@sendgrid/mail');
  const { SENDGRID_API_KEY, SENDGRID_FROM } = process.env;
  if (!SENDGRID_API_KEY || !SENDGRID_FROM) {
    return res.status(500).json({ error: 'SENDGRID_API_KEY or SENDGRID_FROM not set' });
  }
  sgMail.setApiKey(SENDGRID_API_KEY);
  try {
    await sgMail.send({
      from: SENDGRID_FROM,
      to: SENDGRID_FROM,
      subject: 'SendGrid email test',
      text: 'If you see this, SendGrid email is working.',
    });
    res.json({ success: true, from: SENDGRID_FROM });
  } catch (err) {
    res.status(500).json({ error: err.message, code: err.code, response: err.response?.body });
  }
});

// Public routes
app.use('/api/auth',  require('./modules/auth/auth.routes'));
app.use('/api/desk',  require('./modules/kiosk/kiosk.routes'));

// Protected routes
app.use('/api/companies',      authMiddleware, authMiddleware.requireAdmin, require('./modules/companies/company.routes'));
app.use('/api/departments',    authMiddleware, authMiddleware.requireAdmin, require('./modules/companies/department.routes'));
app.use('/api/units',          authMiddleware, authMiddleware.requireAdmin, require('./modules/companies/unit.routes'));
app.use('/api/employees',      authMiddleware, authMiddleware.requireAdmin, require('./modules/employees/employee.routes'));
app.use('/api/attendance',     authMiddleware, authMiddleware.requireAdmin, require('./modules/attendance/attendance.routes'));
app.use('/api/settings',       authMiddleware, authMiddleware.requireAdmin, require('./modules/settings/settings.routes'));
app.use('/api/reports',        authMiddleware, authMiddleware.requireAdmin, require('./modules/reports/report.routes'));
app.use('/api/dashboard',      authMiddleware, authMiddleware.requireAdmin, require('./modules/dashboard/dashboard.routes'));
app.use('/api/access',         authMiddleware, authMiddleware.requireAdmin, authMiddleware.requireSuperAdmin, require('./modules/adminAccounts/adminAccounts.routes'));
app.use('/api/analytics',      authMiddleware, authMiddleware.requireAdmin, require('./modules/analytics/analytics.routes'));

// Swagger UI - API documentation
app.use('/api/docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));

app.use((err, req, res, next) => {
  console.error(err.stack);
  res.status(err.status || 500).json({ error: err.message || 'Internal server error' });
});

const PORT = process.env.PORT || 5000;

initializeDatabase()
  .then(() => {
    app.listen(PORT, () => {
      console.log(`Server running on http://localhost:${PORT}`);
      console.log(`Environment: ${process.env.NODE_ENV || 'development'}`);
    });
  })
  .catch(err => {
    console.error('Failed to initialize database:', err.message);
    console.error('Check your PostgreSQL credentials in .env');
    process.exit(1);
  });