# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Attendance System API",
    "version": "1.0.0",
    "description": "REST API for the Attendance System. Most endpoints require a Bearer JWT token — use POST /api/auth/login to get one, then click the Authorize button at the top right and paste the token."
  },
  "servers": [
    { "url": "https://jamfortetech.com/verifyin-api", "description": "Production (cPanel)" },
    { "url": "http://localhost:5000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CompanyRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "example": "Acme Corp" },
          "address": { "type": "string", "example": "123 Main St, Lagos" },
          "latitude": { "type": "number", "example": 6.5244 },
          "longitude": { "type": "number", "example": 3.3792 },
          "radius_meters": { "type": "integer", "example": 100 },
          "default_shift_start": { "type": "string", "example": "08:00" },
          "default_shift_end": { "type": "string", "example": "17:00" }
        }
      },
...
      "DepartmentRequest": {
        "type": "object",
        "required": ["name", "company_id"],
        "properties": {
          "name": { "type": "string", "example": "Engineering" },
          "company_id": { "type": "integer", "example": 1 }
        }
      },
...
      "UnitRequest": {
        "type": "object",
        "required": ["name", "department_id"],
        "properties": {
          "name": { "type": "string", "example": "Frontend Team" },
          "department_id": { "type": "integer", "example": 2 }
        }
      },
...
      "EmployeeRequest": {
        "type": "object",
        "required": ["name", "company_id"],
        "properties": {
          "employee_id": { "type": "string", "example": "EMP-001" },
          "name": { "type": "string", "example": "Jane Doe" },
          "email": { "type": "string", "example": "jane@company.com" },
          "phone": { "type": "string", "example": "+2348012345678" },
          "company_id": { "type": "integer", "example": 1 },
          "department_id": { "type": "integer", "example": 2 },
          "unit_id": { "type": "integer", "example": 3 },
          "shift_start": { "type": "string", "example": "08:00" },
          "shift_end": { "type": "string", "example": "17:00" },
          "work_days": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday"]
          }
        }
      },
...
      "AdminAccountRequest": {
        "type": "object",
        "required": ["username", "company_ids"],
        "properties": {
          "username": { "type": "string", "example": "branchmanager" },
          "company_ids": {
            "type": "array",
            "items": { "type": "integer" },
            "example": [1, 2]
          }
        }
      },
      "SettingsUpdateRequest": {
        "type": "object",
        "properties": {
          "company_name": { "type": "string", "example": "Acme Corp" },
          "kiosk_greeting": { "type": "string", "example": "Welcome!" },
          "gps_auto_update": { "type": "boolean", "example": true }
        }
      }
    }
  },
  "paths": {
...
    "/api/desk/departments": {
      "get": {
        "tags": ["Kiosk"],
        "summary": "List departments by company (public)",
        "parameters": [
          { "name": "company_id", "in": "query", "required": false, "schema": { "type": "integer" }, "example": 1 }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/desk/units": {
      "get": {
        "tags": ["Kiosk"],
        "summary": "List units by department (public)",
        "parameters": [
          { "name": "department_id", "in": "query", "required": false, "schema": { "type": "integer" }, "example": 2 }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/desk/employees": {
      "get": {
        "tags": ["Kiosk"],
        "summary": "List employees (public)",
        "parameters": [
          { "name": "company_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "department_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "unit_id", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/companies": {
      "get": {
        "tags": ["Companies"],
        "summary": "List all companies",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["Companies"],
        "summary": "Create a company",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompanyRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Created" } }
      }
    },
...
    "/api/departments": {
      "get": {
        "tags": ["Departments"],
        "summary": "List departments (optionally filter by company)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "company_id", "in": "query", "schema": { "type": "integer" }, "example": 1 }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["Departments"],
        "summary": "Create a department",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DepartmentRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Created" } }
      }
    },
...
    "/api/units": {
      "get": {
        "tags": ["Units"],
        "summary": "List units (optionally filter by department)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "department_id", "in": "query", "schema": { "type": "integer" }, "example": 2 }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["Units"],
        "summary": "Create a unit",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UnitRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Created" } }
      }
    },
...
    "/api/employees": {
      "get": {
        "tags": ["Employees"],
        "summary": "List employees with optional filters",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "company_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "department_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "unit_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["active", "inactive"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "example": "Jane" }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["Employees"],
        "summary": "Create an employee",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmployeeRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Created" } }
      }
    },
...
    "/api/attendance": {
      "get": {
        "tags": ["Attendance"],
        "summary": "List attendance logs with optional filters",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "employee_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "date", "in": "query", "schema": { "type": "string", "format": "date" }, "example": "2026-06-08" },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" }, "example": "2026-06-01" },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" }, "example": "2026-06-08" },
          { "name": "department_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "unit_id", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string", "enum": ["clock_in", "clock_out"] }
          },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "example": 50 }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get all system settings",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update system settings",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SettingsUpdateRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Updated" } }
      }
    },
...
    "/api/access": {
      "get": {
        "tags": ["AdminAccounts"],
        "summary": "List all admin accounts (super admin only)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["AdminAccounts"],
        "summary": "Create an admin account (super admin only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminAccountRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Created" } }
      }
    },
...
    "/api/analytics": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get attendance analytics",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "employee_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "department_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "unit_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "date_from", "in": "query", "schema": { "type": "string", "format": "date" }, "example": "2026-06-01" },
          { "name": "date_to", "in": "query", "schema": { "type": "string", "format": "date" }, "example": "2026-06-08" }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    }
  }
}